# src/docs/swagger.yaml
definitions:
...
  dto.CreateGearBoxRequest:
    properties:
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  dto.CreateUpdateCountryDTO:
    properties:
      name:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - name
    type: object
  dto.Filter:
    properties:
      filter_type:
        type: string
      from:
        type: string
      to:
        type: string
      type:
        type: string
    type: object
  dto.GearBoxResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/dto.Filter'
        type: object
      page_number:
        type: integer
      page_size:
        type: integer
      sort:
        items:
          $ref: '#/definitions/dto.SortDTO'
        type: array
    type: object
...
  dto.SortDTO:
    properties:
      col_id:
        type: string
      sort:
        type: string
    type: object
...
  dto.UpdateGearBoxRequest:
    properties:
      name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  helper.Response:
    properties:
      error: {}
      result: {}
      result_code:
        type: integer
      success:
        type: boolean
      validation_error:
        items:
          $ref: '#/definitions/validators.ValidationError'
        type: array
    type: object
  validators.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/country/create:
    post:
      consumes:
      - application/json
      description: Create a Country
      parameters:
      - description: Create a Country
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUpdateCountryDTO'
      responses:
        "201":
          description: Country response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Create a Country
      tags:
      - Countries
  /v1/country/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Country response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Delete a Country
      tags:
      - Countries
  /v1/country/filter:
    post:
      consumes:
      - application/json
      description: GetCountryByFilter
      parameters:
      - description: GetCountryByFilter
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.PaginationInputWithFilter'
      responses:
        "200":
          description: Country response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: GetCountryByFilter
      tags:
      - Countries
  /v1/country/get/{id}:
    get:
      consumes:
      - application/json
      description: Get a Country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Country response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get a Country
      tags:
      - Countries
  /v1/country/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a Country
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Create a Country
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUpdateCountryDTO'
      responses:
        "200":
          description: Country response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Update a Country
      tags:
      - Countries
...
  /v1/gearbox/create:
    post:
      consumes:
      - application/json
      description: Create a GearBox
      parameters:
      - description: Create a GearBox
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGearBoxRequest'
      responses:
        "201":
          description: GearBox response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.GearBoxResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Create a GearBox
      tags:
      - GearBox
  /v1/gearbox/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Get a GearBox
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: GearBox response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get a GearBox
      tags:
      - GearBox
  /v1/gearbox/filter:
    post:
      consumes:
      - application/json
      description: Get GearBox
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.PaginationInputWithFilter'
      responses:
        "200":
          description: gearbox response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Get GearBox
      tags:
      - GearBox
  /v1/gearbox/get/{id}:
    get:
      consumes:
      - application/json
      description: Delete a GearBox
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Delete a GearBox
      tags:
      - GearBox
  /v1/gearbox/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a GearBox
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a GearBox
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateGearBoxRequest'
      responses:
        "200":
          description: GearBox response
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                result:
                  $ref: '#/definitions/dto.GearBoxResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - AuthBearer: []
      summary: Update a GearBox
      tags:
      - GearBox
...
swagger: "2.0"